Issue #49: Redirect sync moved under Sync namespace.
